Make checkboxes conditional upon each other's checked status
This commit provides a conditional nature to the checkboxes. If the user
clicks each of the individual meal type checkboxes (side, main dish, and
desserts) then those three checkboxes uncheck and the entire meal
checkbox is checked FOR the user. If the entire meal checkbox is checked
and the user checks on of the other three, individual checkboxes, the
entire meal check box is unchecked.
This prevents the user from submitting the form with both the entire
meal checkbox checked and any combination of the individual checkboxes
checked as well. It also provides a smoother user experience, by
automatically checking the entire meal checkbox for the user if they
selected all of the other three types.