Loop through array and check for duplicates
Since I modified the structure of the meal object, I implemented a for
loop to iterate over the favorite meals array and return if it found a
match, preventing the user from adding duplicates to the array.

I also made a slight change to the view favorites button click event
listener. Instead of rendering the favorite meals in the toggle favorites
function, the render function is called in the click event listener,
only updating the DOM with the current state of the favorite meals array
when the user clicks to view those favorites.